moving to statefulprocessorhandleimpl
